requirements.txt
google-generativeai
Pillow
gtts
streamlit
python-dotenv